improve html/css
the table of contents is now on the right, taking up as much space
as there is left, and displaying the sections/functions in columns.
this should make it much easier to quickly find the section/function
you are looking for.

add meta tag for mobile devices so they actually benefit from
bootstraps columns.